refactor, satisfy mypy
