+++ crates/python/.stubtest-allowlist
@@ -1,1 +1,2 @@
 qcs_sdk.qcs_sdk
+qcs_sdk._gather_diagnostics
