+++ crates/python/tests/diagnostics.py
@@ -1,0 +1,8 @@
+from qcs_sdk import gather_diagnostics
+
+
+def test_gather_diagnostics():
+    """
+    Assert that gathering diagnostics doesn't panic.
+    """
+    gather_diagnostics()
